Construct URLs with url_for instead of requiring baseURL
Flask provides the url_for helper to construct URLs to endpoint
functions. Any arguments that aren't handled by the method are appended
as query arguments. An absolute URL is constructed when _external is
True. The server address and protocol are properly detected from the
current request. This includes a proxied request so long as the reverse
proxy properly forwards the Host and X-Forwarded-Proto headers. After
this baseURL is not needed, which simplifies the configuration by taking
the hostname out of the picture.