Construct URLs with url_for instead of requiring baseURL
Flask provides the url_for helper to construct URLs to endpoint
functions. Any arguments that aren't handled by the method are appended
as query arguments. An absolute URL is constructed when _external is
True, and the server address is properly detected from the Host header.
After this baseURL is not needed, which simplifies the configuration by
taking the hostname out of the picture.